% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor_distinction.R
\name{actor_distinction}
\alias{actor_distinction}
\title{actor_distinction}
\usage{
actor_distinction(mentions)
}
\arguments{
\item{mentions}{data.frame of mentions}
}
\description{
actor_distinction
}
