% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmetric_wos_mentions.R
\name{altmetric_wos_mentions}
\alias{altmetric_wos_mentions}
\title{altmetric_wos_mentions}
\usage{
altmetric_wos_mentions(mentions, wos_incites, filter = NULL,
  only_tweets = FALSE, retweets = NULL)
}
\arguments{
\item{mentions}{Altmetric.com mentions data.frame}

\item{wos_incites}{InCites WoS records data.frame}

\item{filter}{Character or vector of optional Altmetric.com mention types}

\item{only_tweets}{Boolean indicating if want to reduce Twitter mentions to only tweets}

\item{retweets}{Tweet or retweet data.frame}
}
\description{
This function joins Altmetric mentions with InCites WoS records, optionally can also filter by mention types and remove retweets. As result, it returns mentions with accession number and title from InCites.
}
