% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_authors.R
\name{co_authors}
\alias{co_authors}
\title{co_authors}
\usage{
co_authors(mentions, min = 1, social_media = NULL)
}
\arguments{
\item{mentions}{data.frame of mentions}

\item{min}{Minimun co-occurrences}

\item{social_media}{Vector of social media names}
}
\description{
co_authors
}
