% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_authors_network.R
\name{co_authors_network}
\alias{co_authors_network}
\title{co_authors_network}
\usage{
co_authors_network(mentions, social_media = NULL)
}
\arguments{
\item{mentions}{data.frame of mentions}

\item{social_media}{Vector of social media names}
}
\description{
co_authors_network
}
