% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_keywords_network.R
\name{co_kewords_network}
\alias{co_kewords_network}
\title{co_kewords_network}
\usage{
co_kewords_network(keywords, keywords_mentions, binary = TRUE,
  min_co = 1)
}
\arguments{
\item{keywords}{keywords data.frame}

\item{keywords_mentions}{Altmetric.com mentions with keywords}

\item{binary}{binary counting}

\item{min_co}{minimum of co-occurrences}
}
\description{
co_kewords_network
}
