% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_words.R
\name{co_words}
\alias{co_words}
\title{co_words}
\usage{
co_words(title_words, min_oc = 1, min_co = 1)
}
\arguments{
\item{title_words}{List generated in words extraction functions}

\item{min_oc}{Minimum occurrence}

\item{min_co}{Minimum co-occurrence}
}
\description{
co_words
}
