% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_words_annotate.R
\name{co_words_annotate}
\alias{co_words_annotate}
\title{co_words_annotatde}
\usage{
co_words_annotate(publications_titles)
}
\arguments{
\item{publications_titles}{titles of publications}
}
\description{
co_words_annotatde
}
