% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_words_keywords.R
\name{co_words_keywords}
\alias{co_words_keywords}
\title{co_words_keywords}
\usage{
co_words_keywords(annotate_titles, binary = TRUE)
}
\arguments{
\item{annotate_titles}{list with titles and words annotations}

\item{binary}{if binary counting TRUE}
}
\description{
co_words_keywords
}
