% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_words_tags.R
\name{co_words_tags}
\alias{co_words_tags}
\title{co_words_tags}
\usage{
co_words_tags(annotate_titles)
}
\arguments{
\item{annotate_titles}{list with titles and words annotations}
}
\description{
co_words_tags
}
