% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinction.R
\name{distinction}
\alias{distinction}
\title{distinction}
\usage{
distinction(mentions, type)
}
\arguments{
\item{mentions}{data.frame of mentions}

\item{type}{1 for social media actors and social media types, 2 for research output titles and page URL}
}
\description{
distinction
}
