% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giant_component.R
\name{giant_component}
\alias{giant_component}
\title{giant_component}
\usage{
giant_component(graph)
}
\arguments{
\item{graph}{A igraph graph}
}
\description{
This function returns the main component from a igraph graph.
}
