% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_retuit.R
\name{is_retuit}
\alias{is_retuit}
\title{is_retuit}
\usage{
is_retuit(tweets, consumer_key, consumer_secret, access_token,
  access_secret)
}
\arguments{
\item{tweets}{twitter status ids vector}

\item{consumer_key}{consumer key for Twitter API}

\item{consumer_secret}{consumer secret for Twitter API}

\item{access_token}{access token for Twitter API}

\item{access_secret}{access secret for Twitter API}
}
\description{
is_retuit
}
