% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{network}
\usage{
network(nodes, edges, min = 1, co_min = 1)
}
\arguments{
\item{nodes}{data.frame of nodes}

\item{edges}{data.frame of edges}

\item{min}{Minimun mentions}

\item{co_min}{Minimun co-mentions}
}
\description{
network
}
