% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{profiles}
\usage{
profiles(clusters, legend, keywords_mentions, nsets = 6,
  nintersects = NA)
}
\arguments{
\item{clusters}{Clusters data.frame}

\item{legend}{Ordered vector with the names of the clusters}

\item{keywords_mentions}{Altmetric.com mentions with keywords from altmetric_keywords}

\item{nsets}{Number of sets}

\item{nintersects}{Number of intersects}
}
\description{
profiles
}
