% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socio_semantic_network.R
\name{socio_semantic_network}
\alias{socio_semantic_network}
\title{socio_semantic_network}
\usage{
socio_semantic_network(co_authors_edges, keywords_mentions_cluster,
  layout = 1, legend, nodes_size = c(1:10), label_ratio = 0.1,
  custom_colors_nodes = NULL, custom_colors_clouds = NULL)
}
\arguments{
\item{co_authors_edges}{Edges of co_authors from co_authors_network}

\item{keywords_mentions_cluster}{Altmetric.com mentions with keywords data.frame from co_authors_network}

\item{layout}{Name of layout algorithm (1-Auto, 2-Kamada-Kawai, 3-Fruchterman-Reingold, 4-DrL, 5-Nicely, 6-Components)}

\item{legend}{Ordered vector with the names of the clusters}

\item{nodes_size}{Vector with minimum and maximum value of the nodes}

\item{label_ratio}{Ratio of the size of the node labels}

\item{custom_colors_nodes}{Vector of colors for the nodes}

\item{custom_colors_clouds}{Vector of colors for the clouds}
}
\description{
This function obtains the socio-semantic network
}
