% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic_clusters_ranked.R
\name{thematic_clusters_ranked}
\alias{thematic_clusters_ranked}
\title{thematic_clusters_ranked}
\usage{
thematic_clusters_ranked(mentions, nodes)
}
\arguments{
\item{mentions}{A data.frame of mentions}

\item{nodes}{A network}
}
\description{
thematic_clusters_ranked
}
