% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_annotate.R
\name{words_annotate}
\alias{words_annotate}
\title{words_annotate}
\usage{
words_annotate(publications_titles)
}
\arguments{
\item{publications_titles}{titles of publications}
}
\description{
words_annotate
}
