% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words_keywords.R
\name{words_keywords}
\alias{words_keywords}
\title{words_keywords}
\usage{
words_keywords(annotate_titles, binary = TRUE)
}
\arguments{
\item{annotate_titles}{list with titles and words annotations}

\item{binary}{if binary counting TRUE}
}
\description{
words_keywords
}
