% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{wos_read}
\alias{wos_read}
\title{read_wos}
\usage{
wos_read(path)
}
\arguments{
\item{path}{directory path with all WoS records in txt format}
}
\description{
This function get a directory with WoS records in txt format, and it joins all of them into one data.frame. As result, it returns a data.frame with all records.
}
