% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmetric_keywords.R
\name{altmetric_keywords}
\alias{altmetric_keywords}
\title{altmetric_keywords}
\usage{
altmetric_keywords(altmetric_wos_mentions, wos_records,
  remove_journals = NULL)
}
\arguments{
\item{altmetric_wos_mentions}{Altmetric.com mentions with WoS data}

\item{wos_records}{WoS records}

\item{remove_journals}{name of journals to remove}
}
\description{
altmetric_keywords
}
