% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmetric_wos_mentions.R
\name{altmetric_wos_mentions}
\alias{altmetric_wos_mentions}
\title{altmetric_wos_mentions}
\usage{
altmetric_wos_mentions(mentions, wos, filter = NULL,
  only_tweets = FALSE, retweets = NULL)
}
\arguments{
\item{mentions}{Altmetric.com mentions}

\item{wos}{InCites WoS records}

\item{filter}{filter by mention types}

\item{only_tweets}{reduce Twitter mentions to only tweets}

\item{retweets}{data.frame with tweets and retweets information}
}
\description{
altmetric_wos_mentions
}
