% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_authors_network.R
\name{co_authors_network}
\alias{co_authors_network}
\title{co_authors_network}
\usage{
co_authors_network(clusters, keywords_mentions, min_co = 1)
}
\arguments{
\item{clusters}{A data.frame with all keywords and their clusters}

\item{keywords_mentions}{Altmetric.com mentions with keywords from altmetric_keywords}

\item{min_co}{To filter the network on the basis of a minimum of co-occurrences}
}
\description{
This function gets the keywords and clusters data.frame and keywords mentions from altmetric_keywords and obtains a co-occurrence network of authors. It returns a list wit three data.frame: two of them correspond to the nodes and edges of the network and other one with mentions data.frame with clusters.
}
